move constructing db object to a separate file
